Add some proper mutexing to track object
Additionally, account for the fact that the track could be closed during
poll() and so the subsequent read cannot be certain to succeed.

Much of the existing structure is in place from when xwax was single
threaded with no SDL interface. There are many abuses of the processor
architecture, and the thread communication design needs to be revisited.