MixVibes V2 vinyl definition
The MixVibes vinyl uses the optimal LFSR for the length required for one
side of vinyl.

The ordering of the LR channels differs to the other vinyls. Currently
it's a requirement to reverse L and R input to use this definition. A
flag is needed to switch phase in the definition.