Filter the sync pitch towards 1.0 where no data is available
In previous revisions we set the sync pitch to 1.0 where no data was
available, but this was removed to increase the accurace of the sync
to the timecode.

The downside was that scratching would feel sloppy because the lack
of timecode values during scratching combined with old/outlier values
for sync_pitch meant alignment was quickly lost.

So we assume that the sync pitch is meant to be 1.0, and where we don't
know timecode values to calculate sync pitch, we creep back to this value
using the defined resister-capacitor (RC) filter constant.