Modify timecoder to measure axis crossings 4 times per cycle
This is quite some change to the timecoder. Now the bitstream is
measured in a 'primary' axis, once per cycle (when the axis is
crossed). Pitch information is read 4 times per cycle, when any
axis is crossed. Previously pitch was only measured when one axis
was crossed, in one direction. The current direction is computed
at each axis crossing.