Use a hash lookup table to convert timecode to absolute position
In all cases, a flat lookup table is quite sparse. A hash lookup is
beneficial for the timecodes with more bits.

Use the slot number to define the position. As long as we push into
the hash in position order (which we do) then the two values are
identical.

The benefits of this change are less for the timecodes which already
use a small lookup table (eg. serato_*) than those which use a very
sparse table (eg. traktor_*).