Use _exit(2) to exit from a vfork() child
According to vfork(2) man page, the child must not call exit(3)
as it shares memory with the calling process.

This fixes a bug where, if execl() fails on the importer it results
in SDL spinning on exit(). This is presumably because SDL's atexit()
handler (or something similar) has already been called in the vfork()
child.