Resample to 44100Hz when using fallback decoder
At the moment the sample rate is fixed internally; incoming audio must
match this.

Other decoders do not support resampling; if a user has audio in other
sample rates they must disable the specific decoders and use the fallback
decoder.

The correct fix is to allow variable sample rates internally, the best
design for this is likely to be to use a WAV file to interchange between
the audio decoders and xwax, and have xwax use the sample rate of the
decoded WAV file.