Lower realtime priority
Help the FreeBSD port by using a value which is within range on
both Linux and FreeBSD.

As long as the value is greater than zero and SCHED_FIFO is used,
we expect xwax to take priority over other tasks on the system.
If finer control is needed (eg. to control xwax compared to another
real-time application) then a configuration parameter or command-line
flag would need to be implemented.