Remove return code from audio stop and start functions
These functions must succeed in some form or another as there is no
handling of any errors. Instead of implementing the detailed error
handling required, assert that this functionality is not available.

This leaves us relying on these functions not to error. In the case
of jack_deactivate() this really shouldn't be allowed to error.

Forcing devices to handle errors means the realtime thread can
be kept bug-free and readable.