Guarantee double buffering for ALSA devices
Sometimes ALSA would select a buffer size only slightly larger than
the period size, resulting in xruns. It is necessary to restrict
the choice of period size to allow at least a double buffer.

This fixes a bug which is clear when using 8ms buffer time on
snd-usb-caiaq driver; the driver would select a period size nearly
as large as the number of frames in the buffer.