Check that the realtime thread successfully initialised
It is easier to have the thread itself raise priority and then inform
back via a variable.

If we try and make rt_start() create the thread and then raise its
priority from externally, then we have to deal with killing the thread
on failure conditions. This is more complicated.