Use a memory barrier when incrementing track length
Ensure that everything is in place when the track length is incremented
by the use of a memory barrier.

For this to produce incorrect audio output or display by a small number
of pixels or samples, the compiler would need to re-order instructions
within the commit_pcm_samples() function.

To actually cause a crash would require re-ordering between calls
to track_access_pcm -> more_space and track_commit, so that
tr->block[n] is accessed before it has been set. This is presumably
why this has not been seen in the wild.