Lock and unlock only once
The manual for SDL_UpdateRect is explicit that this must be not called
while the lock is held -- this refactor is not available to us.

Instead, do the redraw and updates separately, and remove some of the
comments which are not really helpful -- better to see the code clearly
and it explains itself.