Be smarter about colouring BPMs
Convert BPM to a hue in the colour wheel. Give compatible BPMs the
same colour by having them end up at the same hue.

Thanks to the folks on the xwax-devel mailing list for persuading me
to re-visit this :)

Also fixes a bug where BPM of 160.0 results in a commpletely different
colour.