Scale the scope and spinner size
This uses a simple fixed method. It would be better to re-allocate
the array based on the size given in the rectangle.

This is going to be troublesome for the timecode monitor, which will
give problems of thread-safety. Probably this is an indication that
the monitor should not be in timecoder.c itself, or we will just need
to use a spinlock appropriately.