Workaround a bug in Make 3.81
Thanks to Daniel Holbach for tracing the bug, which is specific
to Make 3.81 (and appears to be absent in earlier and later versions.

The problem is that the rule

  tests/ttf:   tests/ttf.o

results in Make 3.81 trying to compile the C file directly to ELF
executable, but without the CFLAGS set.

This is a slightly modified version of the original patch because
it leaves just one line which can be removed at a later date.

Make 3.81 is standard on Ubuntu 12.10, so it's fairly common.