Do not use scanf where the empty string is a possibility
No amount of jiggery to scanf() can persuade it to parse an empty
string. This case is much simpler though, so do a simple string split
on the \t character.

Changes the management of the pointers and memory allocation. Instead
of allocating three strings, we keep the original line allocation and
manipulate it.