Always empty the buffer before returning EOF
Fixes a bug where we return EOF when the end of stream is reached,
but we still have data in the rb buffer.

Actually I knew I hadn't implemented this, but forgot to add a
FIXME and hence forgot to add the code. So of course this still
makes it a bug :)

Thanks to Ewan and Daniel for pointing this out.