Refactoring around optimisation
Some basic profiling on a non FPU processor such as ARM seems to suggest
that the interpolation takes most of the time.

This patch is expected to give a minor performance improvement as the
large number of adds/subtract are now done in integers, but still to
do is the multiplacation by mu, mu^2, mu^3.

Mainly it's here to provide clearer input/output to the interpolation
function.