Do not ever reduce the selector to zero lines
In some edge cases of the layout the caller has decided there is enough
room to draw the library, but then there isn't an zero rows can be
fitted.

This is an indication that the whole layout functionality needs a re-think
in the way it deals with making areas invisible as they become too small.

A possible solution for later is separation into a layout and then a redraw
step, whereas at the moment the two are tightly coupled.

I tested with

  xwax -g /2.4

which prior to this patch crashes instantly.