Honour the system locale
We have limited options for text rendering, so must do a conversion
from the system locale to UTF-8 where necessary.

Deliberately not using TTF_RenderUNICODE and wchar_t conversion,
despite it being simpler;  because the system wchar_t is 32-bit,
but SDL's idea of unicode is 16 bit.

Leave iconv out of the path of the rendering that happens every
frame as these are known non-unicode strings.