Honour the search locale
Keep a cache of any records which needed converting to the ASCII
locale used for searching.

These changes would have to be broadened if somebody wanted to
search in a more complex locale; currently the only characters
allowable are [a-z0-9]. It would need a greater understanding of
how SDL handles the user input.