Transfer of copyright from Yves' contributions
Only a few lines of contribution from Yves remain, but for the
avoidance of any doubt I've asked for a transfer of copyright
for all of Yves xwax contributions -- which they have kindly granted
(dated 5th July 2021). This is with the knowledge that there is
the possibility of a license change.

Thanks Yves for all your contributions!

Message-Id: <2d394f00-55fe-8cb5-36d3-4094e85dda49@gmail.com>