These are immutable strings
String literals cannot be changed; to do so is undefined behaviour.
However, their type is not const as they existed before C had the
"const" keyword.

This affects the Mixxx project, who expect to compile the code using
a C++ compiler with minimal changes, and where string literals are
const char*.

Reported-by: Daniel Schürmann <daschuer@mixxx.org>