Do not use sched_setparam() as it's not implemented on musl
The code uses sched_setparam() to set the priority of the given thread,
as per the documentation of GNU's implmenetation. However the POSIX
specification is that this is for the process, which is both not useful
and not what we want.

Because the musl C library is more strict about implementing the POSIX
specification, it leaves this function unimplemented to avoid clashing
with the GNU implementation. So it's better to avoid these issues by
using the specific call to set a thread.