Increase the default size of UI
When xwax was designed, displays were typically 1024x768 or smaller.
It's time for a new default that's more appropriate to modern screens.

This new default size is approximately the same as /1.6, with some
rounding of font size and spaces for the new size.

To approximate the previous size, specify "--geometry /0.625".